Move the Test type to nfa
