+++ lalrpop/src/lexer/nfa/interpret.rs
@@ -1,7 +1,6 @@
 //! A depth-first interpreter for NFAs.
 
-use lexer::nfa::{NFA, NFAStateIndex, START, StateKind, Noop, Other};
-use lexer::re::Test;
+use lexer::nfa::{NFA, NFAStateIndex, START, StateKind, Noop, Other, Test};
 use std::cmp::max;
 
 /// Interpret `nfa` applied to `test`, returning the longest matching
